refactor: remove help flag and move monika command file
