--- src/commands/index.ts
@@ -1,2 +1,0 @@
-export { help } from './help'
-export { flush } from './flush'
