+++ test/index.test.ts
@@ -26,7 +26,7 @@ import { expect, test } from '@oclif/test'
 import path from 'path'
 import chai from 'chai'
 import spies from 'chai-spies'
-import cmd from '../src'
+import cmd from '../src/commands/monika'
 import sinon from 'sinon'
 import * as IpUtil from '../src/utils/public-ip'
